Move  palette ref from palette mono file.
